Move ConfigOption to cli_util.py
