+++ varats/varats/table/tables.py
@@ -13,16 +13,16 @@ from varats.paper_mgmt.artefacts import Artefact, ArtefactFileInfo
 from varats.ts_utils.artefact_util import (
     CaseStudyConverter,
     ReportTypeConverter,
-    ConfigOption,
-    OptionType,
-    COGetter,
-    COGetterV,
     convert_kwargs,
 )
 from varats.ts_utils.cli_util import (
     make_cli_option,
     add_cli_options,
     CLIOptionTy,
+    ConfigOption,
+    OptionType,
+    COGetter,
+    COGetterV,
     cli_yn_choice,
     convert_value,
 )
